Create a logging system: add a new DataCollection model which is used to track the date of each scraper run and which listings were scraped. Doesn't yet communicate with the WebScraper in terms of the outcome, but this may never be needed so for now, YAGNI.
Create list and detail views for the logs. No update or delete functionality needed, just read.

Add a TESTING variable to the global application object. This is used against several if-checks throughout the code to prevent the web scraper from actually being run unless the code isn't in TESTING mode. This should keep the repetitive requests to the production URLs down, particularly if/when testing is added. Once we move to a production server we can remove the TESTING variable and allow the scraper to run.